#6102: move the regression tests into XmirTest, no class exists just for them
